refactor resultSpec build into data connector
so that it can be used by graphql
