+++ master/buildbot/data/connector.py
@@ -24,6 +24,7 @@ from buildbot.data import base
 from buildbot.data import exceptions
 from buildbot.data import resultspec
 from buildbot.data.types import Entity
+from buildbot.util import bytes2unicode
 from buildbot.util import pathmatch
 from buildbot.util import service
 
@@ -79,6 +80,7 @@ class DataConnector(service.AsyncService):
             if inspect.isclass(obj) and issubclass(obj, base.ResourceType):
                 rtype = obj(self.master)
                 setattr(self.rtypes, rtype.name, rtype)
+                setattr(self.plural_rtypes, rtype.plural, rtype)
 
                 # put its update methods into our 'updates' attribute
                 for name in dir(rtype):
@@ -106,6 +108,7 @@ class DataConnector(service.AsyncService):
     def _setup(self):
         self.updates = Updates()
         self.rtypes = RTypes()
+        self.plural_rtypes = RTypes()
         for moduleName in self.submodules:
             module = reflect.namedModule(moduleName)
             self._scanModule(module)
@@ -120,11 +123,14 @@ class DataConnector(service.AsyncService):
     def getResourceType(self, name):
         return getattr(self.rtypes, name)
 
-    @defer.inlineCallbacks
     def get(self, path, filters=None, fields=None, order=None,
             limit=None, offset=None):
         resultSpec = resultspec.ResultSpec(filters=filters, fields=fields,
                                            order=order, limit=limit, offset=offset)
+        return self.get_with_resultspec(path, resultSpec)
+
+    @defer.inlineCallbacks
+    def get_with_resultspec(self, path, resultSpec):
         endpoint, kwargs = self.getEndpoint(path)
         rv = yield endpoint.get(resultSpec, kwargs)
         if resultSpec:
@@ -205,3 +211,96 @@ class DataConnector(service.AsyncService):
             schema += "}\n"
 
         return schema
+
+    def resultspec_from_jsonapi(self, req_args, entityType, is_collection):
+
+        def checkFields(fields, negOk=False):
+            for field in fields:
+                k = bytes2unicode(field)
+                if k[0] == '-' and negOk:
+                    k = k[1:]
+                if k not in entityType.fieldNames:
+                    raise exceptions.InvalidQueryParameter("no such field '{}'".format(k))
+
+        limit = offset = order = fields = None
+        filters, properties = [], []
+        limit = offset = order = fields = None
+        filters, properties = [], []
+        for arg in req_args:
+            argStr = bytes2unicode(arg)
+            if arg == b'order':
+                order = tuple([bytes2unicode(o) for o in req_args[arg]])
+                checkFields(order, True)
+            elif arg == b'field':
+                fields = req_args[arg]
+                checkFields(fields, False)
+            elif arg == b'limit':
+                try:
+                    limit = int(req_args[arg][0])
+                except Exception as e:
+                    raise exceptions.InvalidQueryParameter('invalid limit') from e
+            elif arg == b'offset':
+                try:
+                    offset = int(req_args[arg][0])
+                except Exception as e:
+                    raise exceptions.InvalidQueryParameter('invalid offset') from e
+            elif arg == b'property':
+                try:
+                    props = []
+                    for v in req_args[arg]:
+                        if not isinstance(v, (bytes, str)):
+                            raise TypeError(
+                                "Invalid type {} for {}".format(type(v), v))
+                        props.append(bytes2unicode(v))
+                except Exception as e:
+                    raise exceptions.InvalidQueryParameter(
+                        'invalid property value for {}'.format(arg)) from e
+                properties.append(resultspec.Property(arg, 'eq', props))
+            elif argStr in entityType.fieldNames:
+                field = entityType.fields[argStr]
+                try:
+                    values = [field.valueFromString(v) for v in req_args[arg]]
+                except Exception as e:
+                    raise exceptions.InvalidQueryParameter(
+                        'invalid filter value for {}'.format(argStr)) from e
+
+                filters.append(resultspec.Filter(argStr, 'eq', values))
+            elif '__' in argStr:
+                field, op = argStr.rsplit('__', 1)
+                args = req_args[arg]
+                operators = (resultspec.Filter.singular_operators
+                             if len(args) == 1
+                             else resultspec.Filter.plural_operators)
+                if op in operators and field in entityType.fieldNames:
+                    fieldType = entityType.fields[field]
+                    try:
+                        values = [fieldType.valueFromString(v)
+                                  for v in req_args[arg]]
+                    except Exception as e:
+                        raise exceptions.InvalidQueryParameter(
+                            'invalid filter value for {}'.format(argStr)) from e
+                    filters.append(resultspec.Filter(field, op, values))
+            else:
+                raise exceptions.InvalidQueryParameter(
+                    "unrecognized query parameter '{}'".format(argStr))
+
+        # if ordering or filtering is on a field that's not in fields, bail out
+        if fields:
+            fields = [bytes2unicode(f) for f in fields]
+            fieldsSet = set(fields)
+            if order and {o.lstrip('-') for o in order} - fieldsSet:
+                raise exceptions.InvalidQueryParameter("cannot order on un-selected fields")
+            for filter in filters:
+                if filter.field not in fieldsSet:
+                    raise exceptions.InvalidQueryParameter("cannot filter on un-selected fields")
+
+        # build the result spec
+        rspec = resultspec.ResultSpec(fields=fields, limit=limit, offset=offset,
+                                      order=order, filters=filters, properties=properties)
+
+        # for singular endpoints, only allow fields
+        if not is_collection:
+            if rspec.filters:
+                raise exceptions.InvalidQueryParameter("this is not a collection")
+
+        return rspec
