+++ master/buildbot/data/exceptions.py
@@ -34,3 +34,7 @@ class InvalidPathError(DataException):
 
 class InvalidControlException(DataException):
     "Action is not supported"
+
+
+class InvalidQueryParameter(DataException):
+    "Query Parameter was invalid"
